Remove old ConsPortfolioModel
Swap ALT for previous ConsPortfolioModel.py.  Examples run as intended.  Remove files that depended on old ConsPortfolioModel.
